V 44
final commit

+++ Pollice/app/src/main/java/com/example/hp/pollice/AdminEditPoliceStation.java
@@ -77,7 +77,9 @@ public class AdminEditPoliceStation extends AppCompatActivity {
         delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                new deletePoliceStation().execute("Delete Station", stationId);
+                if(new PublicClass().checkInternetConnection(AdminEditPoliceStation.this)) {
+                    new deletePoliceStation().execute("Delete Station", stationId);
+                }
             }
         });
